+++ src/Components/Search/Search.scss
@@ -1,189 +1,5 @@
 .search {
   display: flex;
   height: 100vh;
-
   background-color: #fffef2;
-
-  h2 {
-    margin-bottom: 12px;
-    color: #333;
-    font-size: 30px;
-    font-weight: 400;
-    line-height: 120%;
-  }
-
-  h3 {
-    margin-bottom: 12px;
-    color: #333;
-    font-size: 26px;
-    font-weight: 400;
-    line-height: 120%;
-  }
-
-  h4 {
-    margin-bottom: 12px;
-    color: #333;
-    font-size: 24px;
-    font-weight: 400;
-    line-height: 120%;
-  }
-
-  h6 {
-    margin-bottom: 8px;
-    color: #111;
-    font-size: 15px;
-    font-weight: 500;
-    line-height: 120%;
-  }
-
-  p {
-    color: #888;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 120%;
-  }
-
-  span {
-    color: #888;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 120%;
-  }
-
-  .icon20 {
-    width: 20px;
-    height: 20px;
-  }
-
-  .icon16 {
-    width: 16px;
-    height: 16px;
-  }
-
-  .divideDot {
-    width: 4px;
-    height: 4px;
-    margin: 8px;
-    border-radius: 4px;
-    background-color: #333;
-  }
-
-  a {
-    color: #333;
-    text-decoration: none;
-  }
-
-  .inputW240 {
-    width: 240px;
-    height: 36px;
-    padding-top: 60px;
-    padding-bottom: 32px;
-    border-bottom: solid 1px #ccc;
-    border-style: none;
-    outline: none;
-    background-color: rgba(000, 000, 000, 0);
-
-    &::placeholder {
-      color: #999;
-      font-size: 24px;
-      font-weight: 300;
-    }
-  }
-
-  .inputW500 {
-    width: 500px;
-    height: 36px;
-    padding-top: 60px;
-    padding-bottom: 32px;
-    border: none;
-    border-bottom: solid 1px #ccc;
-    outline: none;
-    background-color: rgba(000, 000, 000, 0);
-    color: #333;
-    font-size: 30px;
-    font-weight: 300;
-
-    &::placeholder {
-      color: #666;
-      font-size: 24px;
-      font-weight: 300;
-    }
-  }
-
-  .checkBox {
-    width: 500px;
-    margin-bottom: 32px;
-
-    span {
-      margin-left: 12px;
-    }
-  }
-
-  .infoBox {
-    width: 500px;
-    height: 100px;
-    margin-top: 10px;
-    padding: 10px 20px 10px 10px;
-    overflow-y: scroll;
-    border: solid 1px #666;
-    color: #666;
-    font-size: 14px;
-  }
-
-  .secondaryButton {
-    width: 100%;
-    height: 60px;
-    border: solid 1px #ccc;
-    background-color: rgba(000, 000, 000, 0);
-    color: #333;
-  }
-
-  .searchWrap {
-    width: 33%;
-    height: 100vh;
-    padding: 120px 40px;
-    border-right: 1px solid #ccc;
-
-    img {
-      width: 120px;
-    }
-    .searchBox {
-      position: relative;
-      margin-bottom: 40px;
-
-      .searchInput {
-        width: 100%;
-        height: 60px;
-        padding-top: 60px;
-        padding-right: 60px;
-        padding-bottom: 32px;
-        border: none;
-        border-bottom: solid 1px #ccc;
-        outline: none;
-        background-color: rgba(000, 000, 000, 0);
-        color: #333;
-        font-size: 24px;
-        font-weight: 300;
-      }
-
-      .enterButton {
-        position: absolute;
-        top: 50%;
-        right: 0;
-      }
-    }
-
-    .hotKeyword {
-      .keyword {
-        margin-top: 16px;
-        color: #888;
-        font-size: 24px;
-        font-weight: 400;
-      }
-    }
-  }
-
-  .resultWrap {
-    width: 33%;
-  }
 }
