--- src/Components/Search/SearchList/SearchItem/SearchItem.scss
@@ -1,34 +1,0 @@
-.searchItem {
-  width: 100%;
-
-  .hoverResult {
-    right: 0;
-    width: 34%;
-
-    .itemImage {
-      display: none;
-      &.itemHover {
-        display: block;
-        position: absolute;
-        top: 120px;
-        right: 40px;
-        width: 400px;
-      }
-    }
-
-    .primarySolidButton {
-      &.itemHover {
-        position: absolute;
-        right: 0;
-        bottom: 0;
-        width: 34%;
-        height: 60px;
-        border-style: none;
-        background-color: #333;
-        color: #f8f8f8;
-        line-height: 60px;
-        text-align: center;
-      }
-    }
-  }
-}
