+++ src/pages/ProductList/ProductList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 function ProductList() {
-  return <div />;
+  return <div className="productList" />;
 }
 
 export default ProductList;
