+++ src/styles/mixin.scss
@@ -1,0 +1,83 @@
+@mixin primarySolidButton {
+  width: 100%;
+  height: 60px;
+  margin-bottom: 40px;
+  border-style: none;
+  background-color: #333;
+  color: #f8f8f8;
+}
+
+@mixin h2Style {
+  margin-bottom: 12px;
+  color: #333;
+  font-size: 30px;
+  font-weight: 400;
+  line-height: 120%;
+}
+
+@mixin h3Style {
+  margin-bottom: 12px;
+  color: #333;
+  font-size: 26px;
+  font-weight: 400;
+  line-height: 120%;
+}
+
+@mixin h4Style {
+  margin-bottom: 12px;
+  color: #333;
+  font-size: 24px;
+  font-weight: 400;
+  line-height: 120%;
+}
+
+@mixin h6Style {
+  margin-bottom: 8px;
+  color: #111;
+  font-size: 15px;
+  font-weight: 500;
+  line-height: 120%;
+}
+
+@mixin pStyle {
+  color: #888;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 120%;
+}
+
+@mixin spanStyle {
+  color: #888;
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 120%;
+}
+
+@mixin icon24 {
+  width: 24px;
+  height: 24px;
+  transform: translate(100%, 100%);
+}
+
+@mixin icon20 {
+  width: 20px;
+  height: 20px;
+}
+
+@mixin icon16 {
+  width: 16px;
+  height: 16px;
+}
+
+@mixin divideDotStyle {
+  width: 4px;
+  height: 4px;
+  margin: 8px;
+  border-radius: 4px;
+  background-color: #333;
+}
+
+@mixin aStyle {
+  color: #333;
+  text-decoration: none;
+}
